[AIRFLOW-5298] Move FileToGcs to core
For more information check AIP-21.
